🚚👌 move record utility functions directly in segment collection
